Static cast the timestamp to long for serialization.
This patch static casts the timestamp to long for serialization of
timestamps on systems that that use long long time. That should be good
enough until 2037.

R=roubert@google.com

Review URL: https://codereview.appspot.com/132900045

+++ cpp/src/validating_util.cc
@@ -84,7 +84,8 @@ bool UnwrapHeader(const char* header_prefix,
 void ValidatingUtil::Wrap(time_t timestamp, std::string* data) {
   assert(data != NULL);
   char timestamp_string[2 + 3 * sizeof timestamp];
-  int size = std::sprintf(timestamp_string, "%ld", timestamp);
+  int size =
+      std::sprintf(timestamp_string, "%ld", static_cast<long>(timestamp));
   assert(size > 0);
   assert(size < sizeof timestamp_string);
   (void)size;
